+++ src/index.ts
@@ -11,3 +11,4 @@
 
 export { $onEmit } from "./emitter.js";
 export { $lib } from "./lib.js";
+export { $decorators, getServerName } from "./decorators.js";